Backend/requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
numpy
python-dotenv
requests
chromadb 
groq
python-multipart
passlib[argon2]
python-jose[cryptography]
argon2-cffi

# Docling - Advanced Document Parsing
docling
docling-core
docling-parse

# Optional: For better OCR support (if you have scanned documents)
docling-ibm-models

# Web Search and Scraping - Premium APIs + Hybrid Approach
tavily-python>=0.3.0
exa-py>=1.0.0
trafilatura>=1.6.0
playwright>=1.40.0
beautifulsoup4
lxml
aiohttp
pdf2image
motor
gtts